refactor: fix racingcar package
add package car in racingcar.domain
move Car from racingcar.domain to racingcar.domain.car
move Cars from racingcar.domain to racingcar.domain.car
move Position from racingcar.domain to racingcar.domain.car
rename Main to Application
rename CarName to Name
move Name from racingcar.domain to racingcar.domain.car
