+++ src/main/java/racingcar/Application.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import racingcar.controller.RacingCarGameController;
 
-public class Main {
+public class Application {
     public static void main(String[] args) {
         final RacingCarGameController racingCarGameController = new RacingCarGameController();
         racingCarGameController.playGame();
