+++ src/test/java/racingcar/domain/RandomNumberTest.java
@@ -2,6 +2,7 @@ package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.utils.RandomNumber;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
