+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -1,7 +1,8 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.Car;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
